+++ src/util/random.h
@@ -4,7 +4,7 @@
 #include <datatype.h>
 #include <math.h>
 
-void set_seed(uint32_t seed);
+void set_seed(uint64_t seed);
 float64_t uniform(float64_t lower_bound, float64_t upper_bound);
 float64_t normal(float64_t mean, float64_t variance);
